+++ end_to_end/include/case.hpp
@@ -4,9 +4,11 @@
 
 #pragma once
 
+#include <atomic>
+#include <chrono>
+#include <memory>
 #include <string>
 #include <thread>
-#include <variant>
 #include <vector>
 
 #include "action.hpp"
@@ -21,15 +23,16 @@ class TestCase {
                     const std::vector<std::shared_ptr<TestAction>>& actions)
       : name_(name),
         actions_(actions),
-        kill_reception_(std::make_shared<std::atomic_bool>(false)),
         rx_queue_(std::make_shared<SharedQueue<AstarteMessage>>()) {}
 
-  ~TestCase() {
-    if (thread_ && thread_->joinable()) {
-      *kill_reception_ = true;
-      thread_->join();
-    }
-  }
+  ~TestCase() = default;
+
+  // TestCase is not copyable
+  TestCase(const TestCase&) = delete;
+  auto operator=(const TestCase&) -> TestCase& = delete;
+  // TestCase is movable
+  TestCase(TestCase&&) = default;
+  auto operator=(TestCase&& other) -> TestCase& = default;
 
   void configure_curl(const std::string& appengine_url, const std::string& appengine_token,
                       const std::string& realm, const std::string& device_id) {
@@ -37,17 +40,20 @@ class TestCase {
       action->configure_curl(appengine_url, appengine_token, realm, device_id);
     }
   }
+
   void attach_device(const std::shared_ptr<AstarteDeviceGRPC>& device) {
     for (const auto& action : actions_) {
-      action->attach_device(device, rx_queue_, kill_reception_);
+      action->attach_device(device, rx_queue_);
     }
     device_ = device;
   }
+
   void start() {
     if (!thread_) {
-      thread_ = std::make_shared<std::thread>(&TestCase::reception_handler, this);
+      thread_ = std::make_unique<std::jthread>(&TestCase::reception_handler, this);
     }
   }
+
   void execute() {
     spdlog::info("Executing test case: {}.", name_);
     for (const auto& action : actions_) {
@@ -59,12 +65,11 @@ class TestCase {
   std::vector<std::shared_ptr<TestAction>> actions_;
   std::string name_;
   std::shared_ptr<AstarteDeviceGRPC> device_;
-  std::shared_ptr<std::atomic_bool> kill_reception_;
   std::shared_ptr<SharedQueue<AstarteMessage>> rx_queue_;
-  std::shared_ptr<std::thread> thread_;
+  std::unique_ptr<std::jthread> thread_;
 
-  void reception_handler() {
-    while (!(*kill_reception_)) {
+  void reception_handler(std::stop_token token) {
+    while (!token.stop_requested()) {
       auto incoming = device_->poll_incoming(std::chrono::milliseconds(100));
       if (incoming.has_value()) {
         AstarteMessage msg(incoming.value());