+++ samples/qt/main.cpp
@@ -160,12 +160,12 @@ class AstarteWorker : public QObject {
     std::filesystem::path server_property_interface_file_path =
         interfaces_dir / "org.astarte-platform.cpp.examples.ServerProperty.json";
 
-    device->add_interface_from_file(device_individual_interface_file_path, 0ms);
-    device->add_interface_from_file(server_individual_interface_file_path, 0ms);
-    device->add_interface_from_file(device_property_interface_file_path, 0ms);
-    device->add_interface_from_file(device_aggregated_interface_file_path, 0ms);
-    device->add_interface_from_file(server_aggregated_interface_file_path, 0ms);
-    device->add_interface_from_file(server_property_interface_file_path, 0ms);
+    device->add_interface_from_file(device_individual_interface_file_path);
+    device->add_interface_from_file(server_individual_interface_file_path);
+    device->add_interface_from_file(device_property_interface_file_path);
+    device->add_interface_from_file(device_aggregated_interface_file_path);
+    device->add_interface_from_file(server_aggregated_interface_file_path);
+    device->add_interface_from_file(server_property_interface_file_path);
   }
 };